Time: 0 ms (100%), Space: 44.9 MB (78.67%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.java
@@ -19,10 +19,10 @@ class Solution {
     public boolean isValidBST(TreeNode root) {
         // inorder(root);
         // return ans;
-        return validate(root,Integer.MIN_VALUE,Integer.MAX_VALUE);
+        return validate(root,Long.MIN_VALUE,Long.MAX_VALUE);
     }
 
-    public boolean validate(TreeNode root,int min,int max){
+    public boolean validate(TreeNode root,long min,long max){
         if(root==null) return true;
         if(root.val<=min || root.val>=max) return false;
         boolean left=validate(root.left,min,root.val);
